KIT-391: Move selectors file into same test folder, add retries for Cypress test and update Readme
https://coveord.atlassian.net/browse/KIT-391
