+++ packages/atomic/cypress/integration/search-box-selectors.ts
