--- packages/atomic/cypress/selectors/result-list-selectors.ts
@@ -1,7 +1,0 @@
-export const ResultListSelectors = {
-  component: 'atomic-result-list',
-};
-
-export function generateAliasForResultList() {
-  cy.get(ResultListSelectors.component).shadow().as('resultList');
-}
