+++ packages/atomic/src/pages/test.html
@@ -24,7 +24,7 @@
   </head>
 
   <body>
-    <atomic-search-interface id="search" reflect-state-in-url="true" sample>
+    <atomic-search-interface id="search">
       <h1>Content for Integration Test will be replaced here.</h1>
     </atomic-search-interface>
   </body>
